# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export type ProjectSnapshotScope = { [key: string]: unknown };

export interface ProjectSnapshot {
  sourceId: string;
  clientSourceId: string;
  firstName: string;
  lastName: string;
  phone?: string;
  email?: string;
  addressLine1?: string;
  addressLine2?: string;
  city?: string;
  region?: string;
  postalCode?: string;
  /**
   * @maxLength 20
   * @pattern ^$|^[A-Za-z0-9][A-Za-z0-9-]*$
   */
  jobCode?: string;
  /** @maxLength 150 */
  projectName: string;
  salesperson?: string;
  scope: ProjectSnapshotScope;
}

export type EstimateCalculationLinesItem = { [key: string]: unknown };

export type EstimateCalculationTakeoffItem = { [key: string]: unknown };

export interface EstimateCalculation {
  lines: EstimateCalculationLinesItem[];
  takeoff: EstimateCalculationTakeoffItem[];
  issues: string[];
  assumptions: string[];
  [key: string]: unknown;
}

export interface EstimateTotals {
  directCents: number;
  companyCents: number;
  incidentalsCents: number;
  accidentsCents: number;
  salesCents: number;
  beforeTaxCents: number;
  [key: string]: unknown;
}

export type EstimateDraftSlug =
  (typeof EstimateDraftSlug)[keyof typeof EstimateDraftSlug];

export const EstimateDraftSlug = {
  flooring: "flooring",
  bathroom: "bathroom",
  basement: "basement",
} as const;

export interface EstimateDraft {
  id: string;
  slug: EstimateDraftSlug;
  /** @minimum 1 */
  version: number;
  project: ProjectSnapshot;
  calculation: EstimateCalculation;
  totals: EstimateTotals;
  policyVersion: string;
  rateBookVersion: string;
  updatedAt: string;
  createdAt: string;
}

export interface IssuedQuote {
  id: string;
  draftId: string;
  /** @minimum 1 */
  revision: number;
  /** @minimum 1 */
  draftVersion: number;
  project: ProjectSnapshot;
  calculation: EstimateCalculation;
  totals: EstimateTotals;
  policyVersion: string;
  rateBookVersion: string;
  issuedAt: string;
  issuedBy: string;
}

export type EstimateDraftInputSlug =
  (typeof EstimateDraftInputSlug)[keyof typeof EstimateDraftInputSlug];

export const EstimateDraftInputSlug = {
  flooring: "flooring",
  bathroom: "bathroom",
  basement: "basement",
} as const;

export interface EstimateDraftInput {
  slug: EstimateDraftInputSlug;
  project: ProjectSnapshot;
}

export interface EstimateDraftUpdate {
  /** @minimum 1 */
  expectedVersion: number;
  project: ProjectSnapshot;
}

export interface EstimateIssueInput {
  /** @minimum 1 */
  expectedVersion: number;
}

export interface EstimateDraftList {
  drafts: EstimateDraft[];
}

export interface IssuedQuoteList {
  revisions: IssuedQuote[];
}

export interface IntakeSource {
  /** ISO instant with explicit timezone offset */
  updatedAt: string;
  revision?: string;
}

export type EstimatorProjectDeliveryExpectedDocumentsItem =
  (typeof EstimatorProjectDeliveryExpectedDocumentsItem)[keyof typeof EstimatorProjectDeliveryExpectedDocumentsItem];

export const EstimatorProjectDeliveryExpectedDocumentsItem = {
  proposal: "proposal",
  takeoff: "takeoff",
} as const;

export type EstimatorProjectDeliveryClientAddress = {
  line1?: string;
  line2?: string;
  city?: string;
  state?: string;
  postalCode?: string;
};

export type EstimatorProjectDeliveryClient = {
  /** @minLength 1 */
  firstName: string;
  /** @minLength 1 */
  lastName: string;
  phone?: string;
  email?: string;
  address?: EstimatorProjectDeliveryClientAddress;
};

export type EstimatorProjectDeliveryProject = {
  projectType: string;
  name: string;
  jobCode?: string;
};

export interface EstimatorProjectDelivery {
  sourceId: string;
  clientSourceId: string;
  deliveryId: string;
  source: IntakeSource;
  client: EstimatorProjectDeliveryClient;
  project: EstimatorProjectDeliveryProject;
  expectedDocuments?: EstimatorProjectDeliveryExpectedDocumentsItem[];
}

export type EstimatorDocumentDeliveryDocumentType =
  (typeof EstimatorDocumentDeliveryDocumentType)[keyof typeof EstimatorDocumentDeliveryDocumentType];

export const EstimatorDocumentDeliveryDocumentType = {
  proposal: "proposal",
  takeoff: "takeoff",
} as const;

export type EstimatorDocumentDeliveryDocumentMime =
  (typeof EstimatorDocumentDeliveryDocumentMime)[keyof typeof EstimatorDocumentDeliveryDocumentMime];

export const EstimatorDocumentDeliveryDocumentMime = {
  "application/pdf": "application/pdf",
} as const;

export type EstimatorDocumentDeliveryDocument = {
  type: EstimatorDocumentDeliveryDocumentType;
  originalName: string;
  mime: EstimatorDocumentDeliveryDocumentMime;
  sha256: string;
  /** @maxLength 11184812 */
  contentBase64: string;
};

export interface EstimatorDocumentDelivery {
  sourceId: string;
  deliveryId: string;
  source: IntakeSource;
  document: EstimatorDocumentDeliveryDocument;
}

export interface EstimateDeliverySnapshot {
  identityJson: EstimatorProjectDelivery;
  proposalJson: EstimatorDocumentDelivery;
  takeoffJson: EstimatorDocumentDelivery;
  /** @pattern ^[a-f0-9]{64}$ */
  proposalSha256: string;
  /** @pattern ^[a-f0-9]{64}$ */
  takeoffSha256: string;
}

export interface IntakeStatus {
  configured: boolean;
}

export type IntakeReceiptOutcome =
  (typeof IntakeReceiptOutcome)[keyof typeof IntakeReceiptOutcome];

export const IntakeReceiptOutcome = {
  applied: "applied",
  stale: "stale",
} as const;

export interface IntakeReceipt {
  outcome: IntakeReceiptOutcome;
  detail?: string;
}

export interface MeshProject {
  id: string;
  name: string;
  slug: string;
  color: string;
  sort: number;
}

export type MeshProjectDetailsClientSource =
  (typeof MeshProjectDetailsClientSource)[keyof typeof MeshProjectDetailsClientSource];

export const MeshProjectDetailsClientSource = {
  ledger: "ledger",
  "draft-link": "draft-link",
  missing: "missing",
} as const;

export interface MeshProjectDetails {
  projectId: string;
  /** @nullable */
  jobCode: string | null;
  /** @nullable */
  customerName: string | null;
  /** @nullable */
  customerAddress: string | null;
  clientSource: MeshProjectDetailsClientSource;
}

export interface HealthStatus {
  status: string;
}

export type GetMeshStatus200 = {
  configured: boolean;
};

export type ListMeshProjects200 = {
  projects: MeshProject[];
};